Updated code for asset - mutual fund and stocks

+++ earning/EarningService.js
@@ -1,9 +1,21 @@
+const { saveAsset, updateAsset } = require("../asset/assetService");
 const Earning = require("../model/Earning");
 const EarningType = require("../model/EarningType");
 
-async function createEarning(req, res) {
+async function saveEarning({
+  userId,
+  amount,
+  source,
+  description,
+  date,
+  medium,
+  type,
+  assetType,  
+  quantity,
+  assetId,
+  liabilityId
+}) {
   try {
-    const { userId, amount, source, description, date, medium } = req.body;
     let earning = new Earning({
       userId,
       amount,
@@ -11,8 +23,35 @@ async function createEarning(req, res) {
       description,
       date,
       medium,
+      type,
+      assetId,
+      liabilityId
     });
     await earning.save();
+    if (earning.type == 'asset') {      
+      let amt = amount;
+      if (assetType == 'Stock' || assetType == 'Mutual Fund') {
+        amt = quantity;
+      }
+      console.log(amt + " " + quantity + " ")
+
+      updateAsset({assetId, value: amt});
+    } else if (earning.type == 'liability') {
+      // updateLiability();
+    } 
+
+    return earning._id;
+  } catch (error) {
+    console.log(error);
+  }
+  return null;
+}
+
+async function createEarning(req, res) {
+  try {
+    const { userId, amount, type, source, description, date, medium } =
+      req.body;
+    saveEarning({ userId, amount, type, source, description, date, medium });
     res.json({ message: "Inserted Record" });
   } catch (error) {
     res.status(500).json({ error: "Internal Server Error" });
@@ -21,7 +60,7 @@ async function createEarning(req, res) {
 
 async function getAllEarning(req, res) {
   try {
-    const { userId, limit=10, offset=0, filter={} } = req.body;
+    const { userId, limit = 10, offset = 0, filter = {} } = req.body;
     const pageLimit = parseInt(limit, 10);
     const pageOffset = parseInt(offset, 10);
     const { year, month, query } = filter;
@@ -30,33 +69,38 @@ async function getAllEarning(req, res) {
       userId: userId,
       date: {
         $gte: new Date(`${year}-01-01`),
-        $lt: new Date(`${parseInt(year) + 1}-01-01`)
-      }
+        $lt: new Date(`${parseInt(year) + 1}-01-01`),
+      },
     };
 
     if (month) {
       searchQuery.date = {
         $gte: new Date(`${year}-${month}-01`),
-        $lt: new Date(`${year}-${parseInt(month) + 1}-01`)
+        $lt: new Date(`${year}-${parseInt(month) + 1}-01`),
       };
     }
     if (query) {
       searchQuery.$or = [
-        { description: { $regex: query, $options: "i" } }, 
-        { source: { $regex: query, $options: "i" } } 
+        { description: { $regex: query, $options: "i" } },
+        { source: { $regex: query, $options: "i" } },
       ];
     }
 
     // let query = {userId : userId};
     let earningList = await Earning.find(searchQuery)
-    .skip(pageOffset)
-    .limit(pageLimit)
-    .sort({ date: -1 });
+      .skip(pageOffset)
+      .limit(pageLimit)
+      .sort({ date: -1 });
 
     let count = await Earning.countDocuments(searchQuery);
 
-    res.json({ earningList, count, limit: pageLimit, 
-      offset:pageOffset, totalPages: Math.ceil(count/pageLimit)})
+    res.json({
+      earningList,
+      count,
+      limit: pageLimit,
+      offset: pageOffset,
+      totalPages: Math.ceil(count / pageLimit),
+    });
   } catch (error) {
     res.status(500).json({ error: "Internal Server Error" });
   }
@@ -67,8 +111,13 @@ async function updateEarning(req, res) {}
 async function deleteEarningById(req, res) {
   try {
     const { earningId } = req.body;
-    let isDeleted = await Earning.deleteOne({ _id: earningId });
-    res.json({ isDeleted });
+    let earning = Earning.findById(earningId);
+    if (earning && earning.type == "other") {
+      let isDeleted = await Earning.deleteOne({ _id: earningId });
+      res.json({ isDeleted });
+    } else {
+      res.status(500).json({ error: "Please sell/delete asset" });
+    }
   } catch (error) {
     res.status(500).json({ error: "Internal Server Error" });
   }
@@ -112,6 +161,7 @@ async function deleteEarningTypeById(req, res) {
 }
 
 module.exports = {
+  saveEarning,
   updateEarning,
   createEarning,
   getAllEarning,
